(refactor) esm-patient-chart to organize by functionality; eslint ts files
