--- packages/esm-patient-chart-app/src/config-schemas/core-views.ts
@@ -1,10 +1,0 @@
-export interface DashbardLayoutConfig {
-  columns: number;
-}
-
-export interface DashboardConfig {
-  name: string;
-  slot: string;
-  title: string;
-  config: DashbardLayoutConfig;
-}
