--- packages/esm-patient-chart-app/src/config-schemas/index.ts
@@ -1,2 +1,0 @@
-export * from './core-views';
-export * from './openmrs-esm-patient-chart-schema';
