+++ packages/esm-patient-chart-app/src/loader/loader.component.scss
