+++ packages/esm-patient-chart-app/src/patient-chart/action-menu/action-menu.component.scss
