+++ packages/esm-patient-chart-app/src/patient-chart/chart-review/dashboard-view.scss
