+++ packages/esm-patient-chart-app/src/patient-details-tile/patient-details-tile.component.tsx
