+++ packages/esm-patient-chart-app/src/side-nav/side-menu.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from './side-menu.component.scss';
 import { SideNav, SideNavProps } from 'carbon-components-react';
 import { ExtensionSlot, useLayoutType } from '@openmrs/esm-framework';
-import { useNavMenu } from '../hooks/useNavMenu';
+import { useNavMenu } from './useNavMenu';
 import { isDesktop } from '../utils';
 
 interface SideMenuPanelProps extends SideNavProps {}
