+++ packages/esm-patient-chart-app/src/workspace/workspace-renderer.component.tsx
@@ -3,7 +3,7 @@ import { useLayoutType } from '@openmrs/esm-framework';
 import { OpenWorkspace, useWorkspaceWindowSize, WorkspaceWindowState } from '@openmrs/esm-patient-common-lib';
 import { mountRootParcel } from 'single-spa';
 import Parcel from 'single-spa-react/parcel';
-import Loader from '../ui-components/loader.component';
+import Loader from '../loader/loader.component';
 import styles from './workspace-window.scss';
 
 interface WorkspaceRendererProps {
